fix(#793): keep DenyCode::ALL/index OUT of the frozen kinematics talisman
The prior #793 F6 commit added `DenyCode::ALL` + `index()` INSIDE
`crates/kirra-core/src/kinematics_contract.rs` — but that file is the
byte-frozen kinematics talisman (git blob `ed00f4da…`, pinned by the
safety-case talisman gate + the Kani proof). CI's talisman guard correctly
caught the blob mismatch.

Restore the talisman byte-for-byte (blob back to `ed00f4da…`) and move the two
auxiliary items into a NEW non-frozen module `kirra_core::deny_code_ext`
(a second inherent `impl DenyCode` block — legal within the crate). Behaviour
is identical: `DenyCode::ALL` / `.index()` resolve exactly as before for the
`/metrics` `kirra_actuator_denials_total{code=…}` family; only their file
location changed. Adds a consistency test (ALL↔index agree; reasons distinct).

Talisman guard now passes locally; fmt clean; verifier metrics tests green (14).

Refs #793
