+++ crates/kirra-core/src/lib.rs
@@ -29,6 +29,12 @@ use serde::{Deserialize, Serialize};
 #[rustfmt::skip]
 pub mod kinematics_contract;
 
+/// #793 F6 — non-frozen extensions to `DenyCode` (`ALL` / `index`) used by the
+/// `/metrics` actuator-denial family. Kept OUT of the byte-frozen
+/// `kinematics_contract` talisman (blob `ed00f4da…`) so the pin holds; see the
+/// module doc.
+pub mod deny_code_ext;
+
 /// L3.3 — the governor CONSUMER of the frozen Clause-2 cross-partition contract
 /// (`consume_and_bound` / `GovernorVerdict`): read a coherent snapshot →
 /// `validate` the transport contract → decode the `VehicleCommandPayload` →
